pyproject.toml
[project]
name = "asr-tts-rag-pipeline"
version = "0.1.0"
description = "ASR-RAG-TTS pipeline."
readme = "README.md"
authors = [{ name = "Abhyudit Singh", email = "abhyudit.singh@research.iiit.ac.in" }]
requires-python = ">=3.13"
dependencies = [
    "datasets==3.6.0",
    "faiss-cpu>=1.12.0",
    "hydra-core>=1.3.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "loguru>=0.7.3",
    "openai>=1.106.1",
    "openai-whisper>=20250625",
    "sentence-transformers>=5.1.0",
    "sentencepiece>=0.2.1",
    "soundfile>=0.13.1",
]
